@@ -49,6 +49,8 @@ using System.Windows.Controls.Primitives;
 using System.Runtime.CompilerServices;
 using System.Windows.Interop;
 using System.Windows.Media.Imaging;
+using Underanalyzer.Decompiler;
+using UndertaleModLib.Project;
 
 namespace UndertaleModTool
 {
@@ -62,16 +64,22 @@ namespace UndertaleModTool
 
         public UndertaleData Data { get; set; }
         public string FilePath { get; set; }
+        public ProjectContext Project { get; set; } = null;
         public string ScriptPath { get; set; } // For the scripting interface specifically
 
         public string TitleMain { get; set; }
 
-        public static RoutedUICommand CloseTabCommand = new RoutedUICommand("Close current tab", "CloseTab", typeof(MainWindow));
-        public static RoutedUICommand CloseAllTabsCommand = new RoutedUICommand("Close all tabs", "CloseAllTabs", typeof(MainWindow));
-        public static RoutedUICommand RestoreClosedTabCommand = new RoutedUICommand("Restore last closed tab", "RestoreClosedTab", typeof(MainWindow));
-        public static RoutedUICommand SwitchToNextTabCommand = new RoutedUICommand("Switch to the next tab", "SwitchToNextTab", typeof(MainWindow));
-        public static RoutedUICommand SwitchToPrevTabCommand = new RoutedUICommand("Switch to the previous tab", "SwitchToPrevTab", typeof(MainWindow));
-        public static RoutedUICommand SearchInCodeCommand = new("Search in code", "SearchInCode", typeof(MainWindow));
+        public static readonly RoutedUICommand CloseTabCommand = new RoutedUICommand("Close current tab", "CloseTab", typeof(MainWindow));
+        public static readonly RoutedUICommand CloseAllTabsCommand = new RoutedUICommand("Close all tabs", "CloseAllTabs", typeof(MainWindow));
+        public static readonly RoutedUICommand RestoreClosedTabCommand = new RoutedUICommand("Restore last closed tab", "RestoreClosedTab", typeof(MainWindow));
+        public static readonly RoutedUICommand SwitchToNextTabCommand = new RoutedUICommand("Switch to the next tab", "SwitchToNextTab", typeof(MainWindow));
+        public static readonly RoutedUICommand SwitchToPrevTabCommand = new RoutedUICommand("Switch to the previous tab", "SwitchToPrevTab", typeof(MainWindow));
+        public static readonly RoutedUICommand SearchInCodeCommand = new("Search in code", "SearchInCode", typeof(MainWindow));
+        public static readonly RoutedUICommand NewProjectCommand = new("New project", "NewProject", typeof(MainWindow));
+        public static readonly RoutedUICommand OpenProjectCommand = new("Open project", "OpenProject", typeof(MainWindow));
+        public static readonly RoutedUICommand SaveProjectCommand = new("Save project", "SaveProject", typeof(MainWindow));
+        public static readonly RoutedUICommand ViewProjectAssetsCommand = new("View project assets", "ViewProjectAssets", typeof(MainWindow));
+        public static readonly RoutedUICommand CloseProjectCommand = new("Close project", "CloseProject", typeof(MainWindow));
 
         public ObservableCollection<Tab> Tabs { get; set; } = new();
         public Tab CurrentTab
@@ -96,16 +104,24 @@ namespace UndertaleModTool
                 OpenInTab(value);
             } 
         }
-        
+
+        /// <summary>
+        /// Whether the currently-selected object is exportable to a project.
+        /// </summary>
+        public bool IsSelectedProjectExportable { get; private set; }
+
         private static Visibility ToVisibility(bool visible) => visible ? Visibility.Visible : Visibility.Collapsed;
 
-        public Visibility IsGMS2 => ToVisibility((Data?.GeneralInfo?.Major ?? 0) >= 2);
-        public Visibility IsExtProductIDEligible => ToVisibility(UndertaleExtension.ProductDataEligible(Data));
+        public Visibility IsGMS2 => (Data?.GeneralInfo?.Major ?? 0) >= 2 ? Visibility.Visible : Visibility.Collapsed;
+        // God this is so ugly, if there's a better way, please, put in a pull request
+        public Visibility IsExtProductIDEligible => (((Data?.GeneralInfo?.Major ?? 0) >= 2) || (((Data?.GeneralInfo?.Major ?? 0) == 1) && (((Data?.GeneralInfo?.Build ?? 0) >= 1773) || ((Data?.GeneralInfo?.Build ?? 0) == 1539)))) ? Visibility.Visible : Visibility.Collapsed;
 
         public List<Tab> ClosedTabsHistory { get; } = new();
 
         private List<(GMImage, WeakReference<BitmapSource>)> _bitmapSourceLookup { get; } = new();
         private object _bitmapSourceLookupLock = new();
+
+        private ProjectAssetsWindow _projectAssetsWindow = null;
 
         public bool CanSave { get; set; }
         public bool CanSafelySave = false;
@@ -184,14 +200,8 @@ namespace UndertaleModTool
 
         private LoaderDialog scriptDialog;
 
-        // Related to profile system and appdata
-        public byte[] MD5PreviouslyLoaded = null;
-        public byte[] MD5CurrentlyLoaded = null;
+        // Folder used for configuration, etc.
         public static string AppDataFolder => Settings.AppDataFolder;
-        public static string ProfilesFolder { get; } = Path.Combine(Settings.AppDataFolder, "Profiles");
-        public static string CorrectionsFolder { get; } = Path.Combine(Program.GetExecutableDirectory(), "Corrections");
-        public string ProfileHash = null;
-        public bool CrashedWhileEditing = false;
 
         // Scripting interface-related
         private ScriptOptions scriptOptions;
@@ -225,6 +235,10 @@ namespace UndertaleModTool
             { SystemColors.InactiveSelectionHighlightBrushKey, new SolidColorBrush(Color.FromArgb(255, 112, 112, 112)) }
         };
 
+        // Filters for all data files, and for only main data files
+        private const string DataFileFilter = "GameMaker data files (.win, .unx, .ios, .droid, audiogroup*.dat)|*.win;*.unx;*.ios;*.droid;audiogroup*.dat|All files|*";
+        private const string MainDataFileFilter = "GameMaker main data files (.win, .unx, .ios, .droid)|*.win;*.unx;*.ios;*.droid|All files|*";
+
         public MainWindow()
         {
             InitializeComponent();
@@ -240,18 +254,10 @@ namespace UndertaleModTool
 
             scriptSetupTask = Task.Run(() =>
             {
-                scriptOptions = ScriptOptions.Default
-                                .AddImports("UndertaleModLib", "UndertaleModLib.Models", "UndertaleModLib.Decompiler",
-                                            "UndertaleModLib.Scripting", "UndertaleModLib.Compiler",
-                                            "UndertaleModTool", "System", "System.IO", "System.Collections.Generic",
-                                            "System.Text.RegularExpressions")
-                                .AddReferences(typeof(UndertaleObject).GetTypeInfo().Assembly,
-                                                GetType().GetTypeInfo().Assembly,
-                                                typeof(JsonConvert).GetTypeInfo().Assembly,
-                                                typeof(System.Text.RegularExpressions.Regex).GetTypeInfo().Assembly,
-                                                typeof(ImageMagick.MagickImage).GetTypeInfo().Assembly,
-                                                typeof(Underanalyzer.Decompiler.DecompileContext).Assembly)
-                                .WithEmitDebugInformation(true); //when script throws an exception, add a exception location (line number)
+                scriptOptions = ScriptingUtil.CreateDefaultScriptOptions()
+                                             .AddImports("UndertaleModTool")
+                                             .AddReferences(GetType().GetTypeInfo().Assembly,
+                                                            typeof(JsonConvert).GetTypeInfo().Assembly);
             });
 
             var resources = Application.Current.Resources;
@@ -304,11 +310,6 @@ namespace UndertaleModTool
                 _bitmapSourceLookup.Add((image, new WeakReference<BitmapSource>(bitmap)));
                 return bitmap;
             }
-        }
-
-        private void SetIDString(string str)
-        {
-            ((Label)this.FindName("ObjectLabel")).Content = str;
         }
 
         // "attr" is actually "DwmWindowAttribute", but I only need the one value from it
@@ -509,10 +510,6 @@ namespace UndertaleModTool
                     _ = ListenChildConnection(args[2]);
                 }
             }
-
-            // Copy the known code corrections into the profile, if they don't already exist.
-            ApplyCorrections();
-            CrashCheck();
 
             RunGMSDebuggerItem.Visibility = ToVisibility(Settings.Instance.ShowDebuggerOption);
         }
@@ -741,10 +738,11 @@ namespace UndertaleModTool
 
         public async Task<bool> DoOpenDialog()
         {
-            OpenFileDialog dlg = new OpenFileDialog();
-
-            dlg.DefaultExt = "win";
-            dlg.Filter = "GameMaker data files (.win, .unx, .ios, .droid, audiogroup*.dat)|*.win;*.unx;*.ios;*.droid;audiogroup*.dat|All files|*";
+            OpenFileDialog dlg = new()
+            {
+                DefaultExt = "win",
+                Filter = DataFileFilter
+            };
 
             if (dlg.ShowDialog(this) == true)
             {
@@ -755,12 +753,24 @@ namespace UndertaleModTool
         }
         public async Task<bool> DoSaveDialog(bool suppressDebug = false)
         {
-            SaveFileDialog dlg = new SaveFileDialog();
+            // If a project is open, save to its save data file specifically
+            if (Project is not null)
+            {
+                if (this.ShowQuestionWithCancel("Save to the project's designated data file for saving?") == MessageBoxResult.Yes)
+                {
+                    await SaveFile(Project.SaveDataPath, suppressDebug);
+                    return true;
+                }
+                return false;
+            }
 
-            dlg.DefaultExt = "win";
-            dlg.Filter = "GameMaker data files (.win, .unx, .ios, .droid, audiogroup*.dat)|*.win;*.unx;*.ios;*.droid;audiogroup*.dat|All files|*";
-            dlg.FileName = FilePath;
-
+            // Regular save dialog
+            SaveFileDialog dlg = new()
+            {
+                DefaultExt = "win",
+                Filter = DataFileFilter,
+                FileName = FilePath
+            };
             if (dlg.ShowDialog(this) == true)
             {
                 await SaveFile(dlg.FileName, suppressDebug);
@@ -807,6 +817,15 @@ namespace UndertaleModTool
 
         private void Command_Open(object sender, ExecutedRoutedEventArgs e)
         {
+            if (Project is not null)
+            {
+                if (this.ShowQuestionWithCancel("A project is currently open - open another data file and discard all unsaved changes?", MessageBoxImage.Warning, "Project currently open") != MessageBoxResult.Yes)
+                {
+                    // Abort opening new file
+                    return;
+                }
+            }
+
             _ = DoOpenDialog();
         }
 
@@ -832,7 +851,16 @@ namespace UndertaleModTool
 
                 bool save = false;
 
-                if (SettingsWindow.WarnOnClose)
+                if (Project is not null && Project.HasUnexportedAssets)
+                {
+                    MessageBoxResult result = this.ShowQuestionWithCancel("There are assets marked to be exported in the current project.\nReally quit?", MessageBoxImage.Warning, "Warning");
+
+                    if (result != MessageBoxResult.Yes)
+                    {
+                        return;
+                    }
+                }
+                else if (SettingsWindow.WarnOnClose)
                 {
                     MessageBoxResult result = this.ShowQuestionWithCancel("Save changes before quitting?");
 
@@ -862,12 +890,6 @@ namespace UndertaleModTool
                         return;
                     }
                 }
-                else
-                {
-                    RevertProfile();
-                }
-
-                DestroyUMTLastEdited();
 
                 CloseOtherWindows();
 
@@ -989,6 +1011,9 @@ namespace UndertaleModTool
                 Tabs.Clear();
                 ClosedTabsHistory.Clear();
 
+                // Get rid of project
+                UnloadProject();
+
                 // Update GUI and wait for all background processes to finish
                 UpdateLayout();
                 Dispatcher.Invoke(() => { }, DispatcherPriority.ApplicationIdle);
@@ -1057,7 +1082,7 @@ namespace UndertaleModTool
                     return;
                 }
 
-                Dispatcher.Invoke(async () =>
+                Dispatcher.Invoke(() =>
                 {
                     if (data is not null)
                     {
@@ -1077,7 +1102,6 @@ namespace UndertaleModTool
                         {
                             CanSave = true;
                             CanSafelySave = true;
-                            await UpdateProfile(data, filename);
                             data.ToolInfo.DecompilerSettings = SettingsWindow.DecompilerSettings;
                             data.ToolInfo.InstanceIdPrefix = () => SettingsWindow.InstanceIdPrefix;
                         }
@@ -1138,10 +1162,12 @@ namespace UndertaleModTool
             GC.Collect();
         }
 
-        private async Task SaveFile(string filename, bool suppressDebug = false)
+        private async Task<bool> SaveFile(string filename, bool suppressDebug = false)
         {
             if (Data == null || Data.UnsupportedBytecodeVersion)
-                return;
+            {
+                return false;
+            }
 
             bool isDifferentPath = FilePath != filename;
 
@@ -1158,9 +1184,9 @@ namespace UndertaleModTool
             DebugDataDialog.DebugDataMode debugMode = DebugDataDialog.DebugDataMode.NoDebug;
             if (!suppressDebug && Data.GeneralInfo != null && !Data.GeneralInfo.IsDebuggerDisabled)
                 this.ShowWarning("You are saving the game in GameMaker Studio debug mode. Unless the debugger is running, the normal runtime will simply hang after loading. You can turn this off in General Info by checking the \"Disable Debugger\" box and saving.", "GMS Debugger");
-            Task t = Task.Run(async () =>
+            Task<bool> t = Task.Run(() =>
             {
-                bool SaveSucceeded = true;
+                bool saveSucceeded = true;
 
                 try
                 {
@@ -1264,20 +1290,17 @@ namespace UndertaleModTool
                         this.ShowError("An error occurred while trying to save:\n" + e.Message, "Save error");
                     });
 
-                    SaveSucceeded = false;
+                    saveSucceeded = false;
                 }
                 // Don't make any changes unless the save succeeds.
                 try
                 {
-                    if (SaveSucceeded)
+                    if (saveSucceeded)
                     {
                         // It saved successfully!
                         // If we're overwriting a previously existing data file, we're going to overwrite it now.
                         // Then, we're renaming it back to the proper (non-temp) file name.
                         File.Move(filename + "temp", filename, true);
-
-                        // Also make the changes to the profile system.
-                        await ProfileSaveEvent(Data, filename);
                     }
                     else
                     {
@@ -1285,7 +1308,6 @@ namespace UndertaleModTool
                         // We need to delete the temp file though (if it exists).
                         if (File.Exists(filename + "temp"))
                             File.Delete(filename + "temp");
-                        // No profile system changes, since the save failed, like a save was never attempted.
                     }
                 }
                 catch (Exception exc)
@@ -1295,7 +1317,7 @@ namespace UndertaleModTool
                         this.ShowError("An error occurred while trying to save:\n" + exc.Message, "Save error");
                     });
 
-                    SaveSucceeded = false;
+                    saveSucceeded = false;
                 }
 
                 UndertaleCodeEditor.gettextJSON = null;
@@ -1304,12 +1326,16 @@ namespace UndertaleModTool
                 {
                     dialog.Hide();
                 });
+
+                return saveSucceeded;
             });
             dialog.ShowDialog();
-            await t;
+            bool succeeded = await t;
 
             GCSettings.LargeObjectHeapCompactionMode = GCLargeObjectHeapCompactionMode.CompactOnce;
             GC.Collect();
+
+            return succeeded;
         }
 
         public string GenerateMD5(string filename)
@@ -1537,8 +1563,16 @@ namespace UndertaleModTool
             bool isLast = list.IndexOf(obj) == list.Count - 1;
             if (this.ShowQuestion("Delete " + obj + "?" + (!isLast ? "\n\nNote that the code often references objects by ID, so this operation is likely to break stuff because other items will shift up!" : ""), isLast ? MessageBoxImage.Question : MessageBoxImage.Warning, "Confirmation" ) == MessageBoxResult.Yes)
             {
+                // Remove object from list
                 list.Remove(obj);
 
+                // If we're in a project, and this asset is marked for export, unmark it
+                if (Project is not null && obj is IProjectAsset projectAsset)
+                {
+                    Project.UnmarkAssetForExport(projectAsset);
+                }
+
+                // Close tabs and update tree
                 while (CloseTab(obj)) ;
                 UpdateTree();
 
@@ -1837,6 +1871,9 @@ namespace UndertaleModTool
                         {
                             scriptResource.Code = UndertaleCode.CreateEmptyEntry(Data, $"gml_Script_{newName}");
                         }
+
+                        // If we're in a project, mark code entry for export
+                        Project?.MarkAssetForExport(scriptResource.Code);
                     }
                     else if (obj is UndertaleCode codeResource)
                     {
@@ -1875,6 +1912,13 @@ namespace UndertaleModTool
                 str.Content = "string" + list.Count;
             }
             list.Add(obj);
+
+            // If object is a project asset type, and we're in a project, mark it for export
+            if (Project is not null && obj is IProjectAsset projectAsset)
+            {
+                Project.MarkAssetForExport(projectAsset);
+            }
+
             UpdateTree();
             HighlightObject(obj);
             OpenInTab(obj, true);
@@ -2252,141 +2296,6 @@ namespace UndertaleModTool
             }
         }
 
-        public string ProcessException(in Exception exc)
-        {
-            // Collect all original trace lines that we want to parse
-            List<string> traceLines = new();
-            Dictionary<string, int> exTypesDict = null;
-            if (exc is AggregateException)
-            {
-                List<string> exTypes = new();
-
-                // Collect trace lines of inner exceptions, and track their exception type names
-                foreach (Exception ex in (exc as AggregateException).InnerExceptions)
-                {
-                    traceLines.AddRange(ex.StackTrace.Split(Environment.NewLine));
-                    exTypes.Add(ex.GetType().FullName);
-                }
-
-                // Create a mapping of each exception type to the number of its occurrences
-                if (exTypes.Count > 1)
-                {
-                    exTypesDict = exTypes.GroupBy(x => x)
-                                         .Select(x => new { Name = x.Key, Count = x.Count() })
-                                         .OrderByDescending(x => x.Count)
-                                         .ToDictionary(x => x.Name, x => x.Count);
-                }
-            }
-            else if (exc.InnerException is not null)
-            {
-                // Collect trace lines of single inner exception
-                traceLines.AddRange(exc.InnerException.StackTrace.Split(Environment.NewLine));
-            }
-            traceLines.AddRange(exc.StackTrace.Split(Environment.NewLine));
-
-            // Iterate over all lines in the stack trace, finding their line numbers and file names
-            List<(string SourceFile, int LineNum)> loadedScriptLineNums = new();
-            int expectedNumScriptTraceLines = 0;
-            try
-            {
-                foreach (string traceLine in traceLines)
-                {
-                    // Only handle trace lines that come from a script
-                    if (traceLine.TrimStart()[..13] == "at Submission") 
-                    {
-                        // Add to total count of expected script trace lines
-                        expectedNumScriptTraceLines++;
-
-                        // Get full path of the script file, within the line
-                        string sourceFile = Regex.Match(traceLine, @"(?<=in ).*\.csx(?=:line \d+)").Value;
-                        if (!File.Exists(sourceFile))
-                            continue;
-
-                        // Try to find line number from the line
-                        const string pattern = ":line ";
-                        int linePos = traceLine.IndexOf(pattern);
-                        if (linePos > 0 && int.TryParse(traceLine[(linePos + pattern.Length)..], out int lineNum))
-                        {
-                            loadedScriptLineNums.Add((sourceFile, lineNum));
-                        }
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                string excString = exc.ToString();
-
-                int endOfPrevStack = excString.IndexOf("--- End of stack trace from previous location ---");
-                if (endOfPrevStack != -1)
-                {
-                    // Keep only stack trace of the script
-                    excString = excString[..endOfPrevStack];
-                }
-
-                return $"An error occurred while processing the exception text.\nError message - \"{e.Message}\"\nThe unprocessed text is below.\n\n" + excString;
-            }
-
-            // Generate final exception text to show.
-            // If we found the expected number of script trace lines, then use them; otherwise, use the regular exception text.
-            string excText;
-            if (loadedScriptLineNums.Count == expectedNumScriptTraceLines)
-            {                
-                // Read the code for the files to know what the code line associated with the stack trace is
-                Dictionary<string, List<string>> scriptsCode = new();
-                foreach ((string sourceFile, int _) in loadedScriptLineNums)
-                {
-                    if (!scriptsCode.ContainsKey(sourceFile))
-                    {
-                        string scriptCode = null;
-                        try
-                        {
-                            scriptCode = File.ReadAllText(sourceFile, Encoding.UTF8);
-                        }
-                        catch (Exception e)
-                        {
-                            string excString = exc.ToString();
-
-                            return $"An error occurred while processing the exception text.\nError message - \"{e.Message}\"\nThe unprocessed text is below.\n\n" + excString;
-                        }
-                        scriptsCode.Add(sourceFile, scriptCode.Split('\n').ToList());
-                    }
-                }
-
-                // Generate custom stack trace
-                string excLines = string.Join('\n', loadedScriptLineNums.Select(pair =>
-                {
-                    string scriptName = Path.GetFileName(pair.SourceFile);
-                    string scriptLine = scriptsCode[pair.SourceFile][pair.LineNum - 1]; // - 1 because line numbers start from 1
-                    return $"Line {pair.LineNum} in script {scriptName}: {scriptLine}"; 
-                }));
-
-                if (exTypesDict is not null)
-                {
-                    string exTypesStr = string.Join(",\n", exTypesDict.Select(x => $"{x.Key}{((x.Value > 1) ? " (x" + x.Value + ")" : string.Empty)}"));
-                    excText = $"{exc.GetType().FullName}: One on more errors occurred:\n{exTypesStr}\n\nThe current stacktrace:\n{excLines}";
-                }
-                else
-                {
-                    excText = $"{exc.GetType().FullName}: {exc.Message}\n\nThe current stacktrace:\n{excLines}";
-                }
-            }
-            else
-            {
-                string excString = exc.ToString();
-
-                int endOfPrevStack = excString.IndexOf("--- End of stack trace from previous location ---");
-                if (endOfPrevStack != -1)
-                {
-                    // Keep only stack trace of the script
-                    excString = excString[..endOfPrevStack];
-                }
-
-                excText = excString;
-            }
-
-            return excText;
-        }
-
         public void InitializeProgressDialog(string title, string msg)
         {
             scriptDialog ??= new LoaderDialog(title, msg)
@@ -2444,17 +2353,11 @@ namespace UndertaleModTool
             }
             catch (Exception exc)
             {
-                bool isScriptException = exc.GetType().Name == "ScriptException";
-                string excString = string.Empty;
-
-                if (!isScriptException)
-                    excString = ProcessException(in exc);
-
                 await StopProgressBarUpdater();
 
                 Console.WriteLine(exc.ToString());
                 Dispatcher.Invoke(() => CommandBox.Text = exc.Message);
-                this.ShowError(isScriptException ? exc.Message : excString, "Script error");
+                this.ShowError(ScriptingUtil.PrettifyException(in exc), "Script error");
                 ScriptExecutionSuccess = false;
                 ScriptErrorMessage = exc.Message;
                 ScriptErrorType = "Exception";
@@ -2464,11 +2367,54 @@ namespace UndertaleModTool
             scriptText = null;
         }
 
+        public string GetDecompiledText(string codeName, GlobalDecompileContext context = null, IDecompileSettings settings = null)
+        {
+            return GetDecompiledText(Data.Code.ByName(codeName), context, settings);
+        }
+        public string GetDecompiledText(UndertaleCode code, GlobalDecompileContext context = null, IDecompileSettings settings = null)
+        {
+            if (code.ParentEntry is not null)
+                return $"// This code entry is a reference to an anonymous function within \"{code.ParentEntry.Name.Content}\", decompile that instead.";
+
+            GlobalDecompileContext globalDecompileContext = context is null ? new(Data) : context;
+            try
+            {
+                return code != null
+                    ? new Underanalyzer.Decompiler.DecompileContext(globalDecompileContext, code, settings ?? Data.ToolInfo.DecompilerSettings).DecompileToString()
+                    : "";
+            }
+            catch (Exception e)
+            {
+                return "/*\nDECOMPILER FAILED!\n\n" + e.ToString() + "\n*/";
+            }
+        }
+
+        public string GetDisassemblyText(UndertaleCode code)
+        {
+            if (code.ParentEntry is not null)
+                return $"; This code entry is a reference to an anonymous function within \"{code.ParentEntry.Name.Content}\", disassemble that instead.";
+
+            try
+            {
+                return code != null ? code.Disassemble(Data.Variables, Data.CodeLocals?.For(code)) : "";
+            }
+            catch (Exception e)
+            {
+                return "/*\nDISASSEMBLY FAILED!\n\n" + e.ToString() + "\n*/"; // Please don't
+            }
+        }
+        public string GetDisassemblyText(string codeName)
+        {
+            return GetDisassemblyText(Data.Code.ByName(codeName));
+        }
+
         public string PromptLoadFile(string defaultExt, string filter)
         {
-            OpenFileDialog dlg = new OpenFileDialog();
-            dlg.DefaultExt = defaultExt ?? "win";
-            dlg.Filter = filter ?? "GameMaker data files (.win, .unx, .ios, .droid, audiogroup*.dat)|*.win;*.unx;*.ios;*.droid;audiogroup*.dat|All files|*";
+            OpenFileDialog dlg = new()
+            {
+                DefaultExt = defaultExt ?? "win",
+                Filter = filter ?? DataFileFilter
+            };
             return dlg.ShowDialog() == true ? dlg.FileName : null;
         }
 
@@ -2958,70 +2904,85 @@ namespace UndertaleModTool
 
         private async void Command_Run(object sender, ExecutedRoutedEventArgs e)
         {
-            if (Data == null)
+            if (Data is null || FilePath is null)
             {
                 ScriptError("Nothing to run!");
                 return;
             }
-            if ((!WasWarnedAboutTempRun) && SettingsWindow.TempRunMessageShow)
+
+            // Get expected game EXE name
+            string gameExeName = Data?.GeneralInfo?.FileName?.Content;
+            if (gameExeName is null)
             {
-                ScriptMessage(@"WARNING:
+                ScriptError("Null game executable name or location");
+                return;
+            }
+
+            // Change behavior depending on whether a project is loaded currently
+            string saveDataFilePath, gameExePath;
+            bool saveSucceeded;
+            if (Project is not null)
+            {
+                // Project is loaded - try to find game EXE in save directory
+                saveDataFilePath = Project.SaveDataPath;
+                gameExePath = Path.Combine(Path.GetDirectoryName(saveDataFilePath), $"{gameExeName}.exe");
+                if (!File.Exists(gameExePath))
+                {
+                    ScriptError($"Cannot find game executable path, expected to find it at: {gameExePath}");
+                    return;
+                }
+
+                // Save to the file that was designated for the project
+                saveSucceeded = await SaveFile(saveDataFilePath, false);
+            }
+            else
+            {
+                // No project loaded - warn about temp run not permanently saving anything
+                if (!WasWarnedAboutTempRun && SettingsWindow.TempRunMessageShow)
+                {
+                    ScriptMessage(@"WARNING:
 Temp running the game does not permanently 
 save your changes. Please ""Save"" the game
 to save your changes. Closing UndertaleModTool
 without using the ""Save"" option can
 result in loss of work.");
-                WasWarnedAboutTempRun = true;
-            }
-            bool saveOk = true;
-            string oldFilePath = FilePath;
-            bool oldDisableDebuggerState = true;
-            int oldSteamValue = 0;
-            oldDisableDebuggerState = Data.GeneralInfo.IsDebuggerDisabled;
-            oldSteamValue = Data.GeneralInfo.SteamAppID;
-            Data.GeneralInfo.SteamAppID = 0;
-            Data.GeneralInfo.IsDebuggerDisabled = true;
-            string TempFilesFolder = (oldFilePath != null ? Path.Combine(Path.GetDirectoryName(oldFilePath), "MyMod.temp") : "");
-            await SaveFile(TempFilesFolder, false);
-            Data.GeneralInfo.SteamAppID = oldSteamValue;
-            FilePath = oldFilePath;
-            Data.GeneralInfo.IsDebuggerDisabled = oldDisableDebuggerState;
-            if (TempFilesFolder == null)
-            {
-                this.ShowWarning("Temp folder is null.");
-                return;
-            }
-            else if (saveOk)
-            {
-                string gameExeName = Data?.GeneralInfo?.FileName?.Content;
-                if (gameExeName == null || FilePath == null)
-                {
-                    ScriptError("Null game executable name or location");
-                    return;
+                    WasWarnedAboutTempRun = true;
                 }
-                string gameExePath = System.IO.Path.Combine(System.IO.Path.GetDirectoryName(FilePath), gameExeName + ".exe");
+
+                // Try to find game EXE
+                gameExePath = Path.Combine(Path.GetDirectoryName(FilePath), $"{gameExeName}.exe");
                 if (!File.Exists(gameExePath))
                 {
-                    ScriptError("Cannot find game executable path, expected: " + gameExePath);
+                    ScriptError($"Cannot find game executable path, expected to find it at: {gameExePath}");
                     return;
                 }
-                if (!File.Exists(TempFilesFolder))
+
+                // Disable debugger/Steam, and save to the folder where the game was loaded from
+                bool oldDisableDebuggerState = Data.GeneralInfo.IsDebuggerDisabled;
+                int oldSteamValue = Data.GeneralInfo.SteamAppID;
+                Data.GeneralInfo.SteamAppID = 0;
+                Data.GeneralInfo.IsDebuggerDisabled = true;
+                saveDataFilePath = Path.Combine(Path.GetDirectoryName(FilePath), "mod_temprun.temp");
+                saveSucceeded = await SaveFile(saveDataFilePath, false);
+                Data.GeneralInfo.SteamAppID = oldSteamValue;
+                Data.GeneralInfo.IsDebuggerDisabled = oldDisableDebuggerState;
+            }
+            
+            // Run the game using new data file
+            if (saveSucceeded)
+            {
+                if (!File.Exists(saveDataFilePath))
                 {
-                    ScriptError("Cannot find game path, expected: " + TempFilesFolder);
+                    ScriptError($"Cannot find game path, expected to find it at: {saveDataFilePath}");
                     return;
                 }
-                if (gameExeName != null)
-                    Process.Start(gameExePath, "-game \"" + TempFilesFolder + "\" -debugoutput \"" + Path.ChangeExtension(TempFilesFolder, ".gamelog.txt") + "\"");
+                // TODO: possibly have a setting to add debug output via
+                //          -debugoutput \"{Path.ChangeExtension(saveDataFilePath, ".gamelog.txt")}\"
+                Process.Start(gameExePath, $"-game \"{saveDataFilePath}\"");
             }
-            else if (!saveOk)
+            else
             {
-                this.ShowWarning("Temp save failed, cannot run.");
-                return;
-            }
-            if (File.Exists(TempFilesFolder))
-            {
-                await Task.Delay(3000);
-                //File.Delete(TempFilesFolder);
+                this.ShowWarning("Save failed, cannot run.");
             }
         }
         private async void Command_RunSpecial(object sender, ExecutedRoutedEventArgs e)
@@ -3142,17 +3103,70 @@ result in loss of work.");
 
         public void UpdateObjectLabel(object obj)
         {
-            int foundIndex = obj is UndertaleResource res ? Data.IndexOf(res, false) : -2;
-            string idString;
+            // Try to get index
+            int foundIndex = obj is UndertaleResource res ? Data.IndexOf(res, false) : -1;
 
+			// Determine ID
+			string idString;
             if (foundIndex == -2)
+			{	
                 idString = "None";
+        	}
             else if (foundIndex == -1)
+            {
                 idString = "N/A";
+            }
             else
-                idString = Convert.ToString(foundIndex);
+            {
+                idString = foundIndex.ToString();
+            }
 
-            SetIDString(idString);
+            // Update main label
+            ObjectLabel.Content = $"ID: {idString}";
+
+            // If this object is a project asset (and we're in an open project), then show/update marked for export status
+            if (Project is not null && foundIndex >= 0 && obj is IProjectAsset { ProjectExportable: true } projectAsset)
+            {
+                IsSelectedProjectExportable = true;
+                MarkedForExportGroup.Visibility = Visibility.Visible;
+                MarkedForExportCheckBox.IsChecked = Project.IsAssetMarkedForExport(projectAsset);
+            }
+            else
+            {
+                IsSelectedProjectExportable = false;
+                MarkedForExportGroup.Visibility = Visibility.Collapsed;
+            }
+        }
+
+        private void MarkedForExport_CheckBoxChanged(object sender, RoutedEventArgs e)
+        {
+            // Ensure we actually have a bool value (could be null)
+            if (MarkedForExportCheckBox.IsChecked is not bool isChecked)
+            {
+                return;
+            }
+
+            // Ensure we're actually in a project
+            if (Project is null)
+            {
+                return;
+            }
+
+            // Ensure our current object is actually a project asset
+            if (Selected is not IProjectAsset projectAsset)
+            {
+                return;
+            }
+
+            // Change state in the project
+            if (isChecked)
+            {
+                Project.MarkAssetForExport(projectAsset);
+            }
+            else
+            {
+                Project.UnmarkAssetForExport(projectAsset);
+            }
         }
 
         public void HighlightObject(object obj, bool silent = true)
@@ -3290,7 +3304,7 @@ result in loss of work.");
             OpenFileDialog dlg = new OpenFileDialog();
 
             dlg.DefaultExt = "win";
-            dlg.Filter = "GameMaker data files (.win, .unx, .ios, .droid)|*.win;*.unx;*.ios;*.droid|All files|*";
+            dlg.Filter = MainDataFileFilter;
 
             if (dlg.ShowDialog() == true)
             {
@@ -3762,6 +3776,304 @@ result in loss of work.");
             }
 
             return false;
+        }
+
+        private void UnloadProject()
+        {
+            Project = null;
+            _projectAssetsWindow?.Close();
+            _projectAssetsWindow = null;
+        }
+
+        private void AssignNewProject(ProjectContext project)
+        {
+            UnloadProject();
+
+            Project = project;
+            project.UnexportedAssetsChanged += (sender, args) =>
+            {
+                UpdateObjectLabel(Selected);
+            };
+            UpdateObjectLabel(Selected);
+        }
+
+        private string ChooseProjectSaveFile()
+        {
+            // Choose data file to save project to (when loading or saving in general)
+            SaveFileDialog saveDataDialog = new()
+            {
+                DefaultExt = "win",
+                Filter = MainDataFileFilter,
+                Title = "Choose data file to save project to"
+            };
+            if (saveDataDialog.ShowDialog(this) != true)
+            {
+                return null;
+            }
+
+            // Check if the directories are the same and warn if so (note: not a fully exhaustive check, but decent)
+            string saveFilePath = saveDataDialog.FileName;
+            if (Path.GetFullPath(Path.GetDirectoryName(saveFilePath)).Equals(
+                    Path.GetFullPath(Path.GetDirectoryName(FilePath)), StringComparison.OrdinalIgnoreCase))
+            {
+                if (this.ShowQuestionWithCancel("The save file path was set to the same directory as the loaded file path. This may permanently overwrite external data files. Proceed?", MessageBoxImage.Warning, "Save file in same directory as loaded file") != MessageBoxResult.Yes)
+                {
+                    // Abort
+                    return null;
+                }
+            }
+
+            // Check if the save directory is empty, and warn if so
+            try
+            {
+                if (!Directory.EnumerateFileSystemEntries(Path.GetDirectoryName(saveFilePath)).Any())
+                {
+                    this.ShowWarning("Currently, the save file's directory is empty. You will likely want to copy all other game files to the save directory, so that external assets can be loaded correctly (both in-game and in this tool), and so the game can run from there.");
+                }
+            }
+            catch (Exception)
+            {
+                // Ignore filesystem errors on the above check; we don't really care
+            }
+
+            return saveFilePath;
+        }
+
+        private void Command_NewProject(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (Data is null || FilePath is null)
+            {
+                // No data set; this should be impossible, but abort if this does occur
+                return;
+            }
+            if (Project is not null && Project.HasUnexportedAssets)
+            {
+                if (this.ShowQuestionWithCancel("There are assets marked to be exported in the current project - create a new project and discard all unexported changes?", MessageBoxImage.Warning, "Project already open") != MessageBoxResult.Yes)
+                {
+                    // Abort new project creation
+                    return;
+                }
+            }
+
+            // Ask for save file directory
+            string saveFilePath = ChooseProjectSaveFile();
+            if (saveFilePath is null)
+            {
+                // Save file prompt failed or was cancelled
+                return;
+            }
+
+            // Ask for name
+            string projectName = SimpleTextInput("Choose project name", "Choose a name for the new project", $"{Data.GeneralInfo?.DisplayName?.Content ?? "New"} Mod", false, true);
+            if (projectName is null)
+            {
+                // Name prompt was canceled
+                SetUMTConsoleText("Cancelled new project creation.");
+                return;
+            }
+            projectName = projectName.Trim();
+
+            // Prompt location for project file
+            string directory = PromptChooseDirectory();
+            if (directory is null)
+            {
+                // Directory prompt was canceled
+                SetUMTConsoleText("Cancelled new project creation.");
+                return;
+            }
+
+            // Attempt making project at the specified location (will fail if the folder isn't empty, etc.)
+            ProjectContext newProjectContext;
+            try
+            {
+                newProjectContext = new(Data, FilePath, saveFilePath, Path.Combine(directory, "project.json"), projectName, (f) => Dispatcher.Invoke(f));
+            }
+            catch (ProjectException ex)
+            {
+                this.ShowError(ex.Message, "Failed to create new project");
+                SetUMTConsoleText("Project creation failed.");
+                return;
+            }
+            catch (Exception ex)
+            {
+                this.ShowError($"Error occurred when creating new project:\n{ex}");
+                SetUMTConsoleText("Project creation failed.");
+                return;
+            }
+
+            // Start using new project context
+            AssignNewProject(newProjectContext);
+            SetUMTConsoleText($"Project \"{projectName}\" created successfully!");
+        }
+
+        private async void Command_OpenProject(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (Data is null || FilePath is null)
+            {
+                // No data set; this should be impossible, but abort if this does occur
+                return;
+            }
+            if (Project is not null && Project.HasUnexportedAssets)
+            {
+                if (this.ShowQuestionWithCancel("There are assets marked to be exported in the current project - open another new project and discard all unexported changes?", MessageBoxImage.Warning, "Project already open") != MessageBoxResult.Yes)
+                {
+                    // Abort opening project
+                    return;
+                }
+            }
+
+            // Ask for save file directory
+            string saveFilePath = ChooseProjectSaveFile();
+            if (saveFilePath is null)
+            {
+                // Save file prompt failed or was cancelled
+                return;
+            }
+
+            // Change main file path to the save data file path
+            string loadFilePath = FilePath;
+            FilePath = saveFilePath;
+
+            // Choose project file to open
+            OpenFileDialog openProjectDialog = new()
+            {
+                DefaultExt = "json",
+                Filter = "UndertaleModTool project files (.json)|*.json|All files|*",
+                Title = "Open project file"
+            };
+            if (openProjectDialog.ShowDialog(this) == true)
+            {
+                ProjectContext newProjectContext = null;
+
+                // Attempt loading project from the specific JSON
+                IsEnabled = false;
+                await Task.Run(() =>
+                {
+                    try
+                    {
+                        newProjectContext = ProjectContext.CreateWithDataFilePaths(loadFilePath, saveFilePath, openProjectDialog.FileName);
+                        newProjectContext.Import(Data, null, (f) => Dispatcher.Invoke(f));
+                    }
+                    catch (ProjectException ex)
+                    {
+                        newProjectContext = null;
+                        this.ShowError(ex.Message, "Failed to load project");
+                    }
+                    catch (Exception ex)
+                    {
+                        newProjectContext = null;
+                        this.ShowError($"Error occurred when loading project:\n{ex}");
+                    }
+                });
+                IsEnabled = true;
+
+                // Don't assign new project context if load failed
+                if (newProjectContext is null)
+                {
+                    SetUMTConsoleText("Project failed to open.");
+                    return;
+                }
+
+                // Start using new project context
+                AssignNewProject(newProjectContext);
+                SetUMTConsoleText($"Opened project \"{newProjectContext.Name}\".");
+            }
+        }
+        private async void Command_SaveProject(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (Data is null || FilePath is null)
+            {
+                // No data set; this should be impossible, but abort if this does occur
+                return;
+            }
+            if (Project is null)
+            {
+                // No project set; this should also be impossible
+                return;
+            }
+
+            // Attempt saving project
+            IsEnabled = false;
+            bool success = false;
+            await Task.Run(() =>
+            {
+                try
+                {
+                    Project.Export(true);
+                    success = true;
+                }
+                catch (ProjectException ex)
+                {
+                    this.ShowError(ex.Message, "Failed to save project");
+                }
+                catch (Exception ex)
+                {
+                    this.ShowError($"Error occurred when saving project:\n{ex}");
+                }
+            });
+            IsEnabled = true;
+            SetUMTConsoleText(success ? "Saved project successfully." : "Project failed to save.");
+        }
+
+        private void Command_ViewProjectAssets(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (Data is null || FilePath is null)
+            {
+                // No data set; this should be impossible, but abort if this does occur
+                return;
+            }
+            if (Project is null)
+            {
+                // No project set; this should also be impossible
+                return;
+            }
+
+            // Focus existing window if already there...
+            if (_projectAssetsWindow is not null)
+            {
+                _projectAssetsWindow.Focus();
+                return;
+            }
+
+            // Open project save window to handle the rest
+            _projectAssetsWindow = new()
+            {
+                Owner = this
+            };
+            _projectAssetsWindow.Closed += (sender, args) =>
+            {
+                _projectAssetsWindow = null;
+            };
+            _projectAssetsWindow.Show();
+        }
+
+        private void Command_CloseProject(object sender, ExecutedRoutedEventArgs e)
+        {
+            if (Data is null || FilePath is null)
+            {
+                // No data set; this should be impossible, but abort if this does occur
+                return;
+            }
+            if (Project is null)
+            {
+                // No project set; this should also be impossible
+                return;
+            }
+
+            // Ensure user really wants to do this
+            if (Project is not null && Project.HasUnexportedAssets)
+            {
+                MessageBoxResult result = this.ShowQuestionWithCancel("There are assets marked to be exported in the current project. Really close?");
+
+                if (result != MessageBoxResult.Yes)
+                {
+                    return;
+                }
+            }
+
+            UnloadProject();
+            UpdateObjectLabel(Selected);
+            SetUMTConsoleText("Project closed.");
         }
     }
 
