@@ -1,10 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Globalization;
 using System.IO;
 using System.Text;
 using Underanalyzer;
 using UndertaleModLib.Decompiler;
+using UndertaleModLib.Project;
+using UndertaleModLib.Project.SerializableAssets;
 using UndertaleModLib.Util;
 // ReSharper disable All
 
@@ -888,6 +891,12 @@ public class UndertaleInstruction : UndertaleObject, IGMInstruction
                 if (type1 == DataType.Int32)
                 {
                     reader.Position += 4;
+
+                    // Existence of this argument implies GameMaker 2023.8 or above
+                    if (!reader.undertaleData.IsVersionAtLeast(2023, 8))
+                    {
+                        reader.undertaleData.SetGMS2Version(2023, 8);
+                    }
                 }
                 break;
             }
@@ -1429,7 +1438,7 @@ public static class UndertaleInstructionUtil
 /// A code entry in a data file.
 /// </summary>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
-public class UndertaleCode : UndertaleNamedResource, UndertaleObjectWithBlobs, IDisposable, IGMCode
+public class UndertaleCode : UndertaleNamedResource, UndertaleObjectWithBlobs, IProjectAsset, INotifyPropertyChanged, IDisposable, IGMCode
 {
     /// <summary>
     /// The name of the code entry.
@@ -1481,6 +1490,11 @@ public class UndertaleCode : UndertaleNamedResource, UndertaleObjectWithBlobs, I
     /// Child entries of this code entry, if a root-level (parent) entry; empty if a child entry.
     /// </summary>
     public List<UndertaleCode> ChildEntries { get; set; } = new List<UndertaleCode>();
+
+    /// <inheritdoc />
+#pragma warning disable CS0067 // TODO: remove this suppression once Fody is no longer in use
+    public event PropertyChangedEventHandler PropertyChanged;
+#pragma warning restore CS0067
 
     // Bytecode address to use during (de)serialization, since bytecode can be a separate blob
     private uint _bytecodeAbsoluteAddress;
@@ -1853,7 +1867,24 @@ public class UndertaleCode : UndertaleNamedResource, UndertaleObjectWithBlobs, I
         Name = null;
         _unsupportedBuffer = null;
     }
-    
+
+    /// <inheritdoc/>
+    ISerializableProjectAsset IProjectAsset.GenerateSerializableProjectAsset(ProjectContext projectContext)
+    {
+        SerializableCode serializable = new();
+        serializable.PopulateFromData(projectContext, this);
+        return serializable;
+    }
+
+    /// <inheritdoc/>
+    public string ProjectName => Name?.Content ?? "<unknown name>";
+
+    /// <inheritdoc/>
+    public SerializableAssetType ProjectAssetType => SerializableAssetType.Code;
+
+    /// <inheritdoc/>
+    public bool ProjectExportable => Name?.Content is not null && ParentEntry is null;
+
     // Underanalyzer implementations
     IGMString IGMCode.Name => Name;
     int IGMCode.Length => (int)Length;
