@@ -1,4 +1,6 @@
 ﻿using System;
+using UndertaleModLib.Project;
+using UndertaleModLib.Project.SerializableAssets;
 
 namespace UndertaleModLib.Models;
 
@@ -6,7 +8,7 @@ namespace UndertaleModLib.Models;
 /// A font entry of a data file.
 /// </summary>
 [PropertyChanged.AddINotifyPropertyChangedInterface]
-public class UndertaleFont : UndertaleNamedResource, IDisposable
+public class UndertaleFont : UndertaleNamedResource, IProjectAsset, IDisposable
 {
     /// <summary>
     /// The name of the font.
@@ -401,4 +403,21 @@ public class UndertaleFont : UndertaleNamedResource, IDisposable
         Texture = null;
         Glyphs = new();
     }
+
+    /// <inheritdoc/>
+    ISerializableProjectAsset IProjectAsset.GenerateSerializableProjectAsset(ProjectContext projectContext)
+    {
+        SerializableFont serializable = new();
+        serializable.PopulateFromData(projectContext, this);
+        return serializable;
+    }
+
+    /// <inheritdoc/>
+    public string ProjectName => Name?.Content ?? "<unknown name>";
+
+    /// <inheritdoc/>
+    public SerializableAssetType ProjectAssetType => SerializableAssetType.Font;
+
+    /// <inheritdoc/>
+    public bool ProjectExportable => Name?.Content is not null && Texture is not null;
 }
