@@ -7,6 +7,8 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using UndertaleModLib.Project;
+using UndertaleModLib.Project.SerializableAssets;
 using UndertaleModLib.Util;
 
 namespace UndertaleModLib.Models;
@@ -14,7 +16,7 @@ namespace UndertaleModLib.Models;
 /// <summary>
 /// A room in a data file.
 /// </summary>
-public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDisposable
+public class UndertaleRoom : UndertaleNamedResource, IProjectAsset, INotifyPropertyChanged, IDisposable
 {
     /// <summary>
     /// Certain flags a room can have.
@@ -129,18 +131,26 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     /// <summary>
     /// The width of the room grid in pixels.
     /// </summary>
+    /// <remarks>
+    /// This is an UndertaleModTool-only property; it doesn't exist in game data.
+    /// </remarks>
     public double GridWidth { get => _gridWidth; set { if (value >= 0) _gridWidth = value; } }
 
     /// <summary>
     /// The height of the room grid in pixels.
     /// </summary>
+    /// <remarks>
+    /// This is an UndertaleModTool-only property; it doesn't exist in game data.
+    /// </remarks>
     public double GridHeight { get => _gridHeight; set { if (value >= 0) _gridHeight = value; } }
 
     /// <summary>
     /// The thickness of the room grid in pixels.
     /// </summary>
+    /// <remarks>
+    /// This is an UndertaleModTool-only property; it doesn't exist in game data.
+    /// </remarks>
     public double GridThicknessPx { get; set; } = 1d;
-    private UndertalePointerList<Layer> _layers = new();
 
     /// <summary>
     /// The list of backgrounds this room uses.
@@ -161,6 +171,8 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     /// The list of tiles this room uses.
     /// </summary>
     public UndertalePointerList<Tile> Tiles { get; set; } = new UndertalePointerList<Tile>();
+
+    private UndertalePointerList<Layer> _layers = new();
 
     /// <summary>
     /// List of instance creation order IDs, used for the first room in the room order only, in GameMaker 2024.13 and above.
@@ -595,6 +607,23 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         Caption = null;
         GameObjects = new();
     }
+
+    /// <inheritdoc/>
+    ISerializableProjectAsset IProjectAsset.GenerateSerializableProjectAsset(ProjectContext projectContext)
+    {
+        SerializableRoom serializable = new();
+        serializable.PopulateFromData(projectContext, this);
+        return serializable;
+    }
+
+    /// <inheritdoc/>
+    public string ProjectName => Name?.Content ?? "<unknown name>";
+
+    /// <inheritdoc/>
+    public SerializableAssetType ProjectAssetType => SerializableAssetType.Room;
+
+    /// <inheritdoc/>
+    public bool ProjectExportable => Name?.Content is not null;
 
     /// <summary>
     /// Interface for objects within rooms.
@@ -1385,7 +1414,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     //TODO: everything from here on is mostly gms2 related which i dont have much experience with
     public class Layer : UndertaleObject, INotifyPropertyChanged, IDisposable
     {
-        public interface LayerData : UndertaleObject, IDisposable
+        public interface LayerData : UndertaleObject, INotifyPropertyChanged, IDisposable
         {
         }
 
@@ -1395,6 +1424,9 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         /// <summary>
         /// The room this layer belongs to.
         /// </summary>
+        /// <remarks>
+        /// This is an UndertaleModTool-only property; it doesn't exist in game data.
+        /// </remarks>
         public UndertaleRoom ParentRoom { get => _parentRoom; set { _parentRoom = value; OnPropertyChanged(); UpdateParentRoom(); } }
 
         /// <summary>
@@ -1569,6 +1601,11 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             internal uint[] InstanceIds { get; set; } // 100000, 100001, 100002, 100003 - instance ids from GameObjects list in the room
             public ObservableCollection<GameObject> Instances { get; set; } = new();
 
+            /// <inheritdoc />
+#pragma warning disable CS0067 // TODO: remove this suppression once Fody is no longer in use
+            public event PropertyChangedEventHandler PropertyChanged;
+#pragma warning restore CS0067
+
             public bool AreInstancesUnresolved()
             {
                 return InstanceIds?.Length > 0 && Instances?.Count == 0;
@@ -1614,7 +1651,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             }
         }
 
-        public class LayerTilesData : LayerData, INotifyPropertyChanged
+        public class LayerTilesData : LayerData
         {
             private UndertaleResourceById<UndertaleBackground, UndertaleChunkBGND> _background = new(); // In GMS2 backgrounds are just tilesets
             private uint _tilesX;
@@ -1954,7 +1991,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             }
         }
 
-        public class LayerBackgroundData : LayerData, IStaticChildObjCount, IStaticChildObjectsSize, INotifyPropertyChanged
+        public class LayerBackgroundData : LayerData, IStaticChildObjCount, IStaticChildObjectsSize
         {
             /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
             public static readonly uint ChildObjectCount = 1;
@@ -1975,7 +2012,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
 
             public bool Visible { get; set; } = true;
             public bool Foreground { get; set; }
-            public UndertaleSprite Sprite { get => _sprite.Resource; set { _sprite.Resource = value; OnPropertyChanged(); ParentLayer.ParentRoom.UpdateBGColorLayer(); } }
+            public UndertaleSprite Sprite { get => _sprite.Resource; set { _sprite.Resource = value; OnPropertyChanged(); ParentLayer?.ParentRoom?.UpdateBGColorLayer(); } }
             public bool TiledHorizontally { get => _tiledHorizontally; set { _tiledHorizontally = value; OnPropertyChanged(); } }
             public bool TiledVertically { get => _tiledVertically; set { _tiledVertically = value; OnPropertyChanged(); } }
             public bool Stretch { get => _stretch; set { _stretch = value; OnPropertyChanged(); } }
@@ -2064,6 +2101,11 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             public UndertalePointerList<SpriteInstance> NineSlices { get; set; } // Removed in 2.3.2, before never used
             public UndertalePointerList<ParticleSystemInstance> ParticleSystems { get; set; }
             public UndertalePointerList<TextItemInstance> TextItems { get; set; }
+
+            /// <inheritdoc />
+#pragma warning disable CS0067 // TODO: remove this suppression once Fody is no longer in use
+            public event PropertyChangedEventHandler PropertyChanged;
+#pragma warning restore CS0067
 
             /// <inheritdoc />
             public void Serialize(UndertaleWriter writer)
@@ -2217,8 +2259,13 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
         [PropertyChanged.AddINotifyPropertyChangedInterface]
         public class LayerEffectData : LayerData
         {
-            public UndertaleString EffectType;
-            public UndertaleSimpleList<EffectProperty> Properties;
+            public UndertaleString EffectType { get; set; }
+            public UndertaleSimpleList<EffectProperty> Properties { get; set; }
+
+            /// <inheritdoc />
+#pragma warning disable CS0067 // TODO: remove this suppression once Fody is no longer in use
+            public event PropertyChangedEventHandler PropertyChanged;
+#pragma warning restore CS0067
 
             /// <inheritdoc />
             public void Serialize(UndertaleWriter writer)
@@ -2775,7 +2822,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
     /// </summary>
     public class InstanceIDList : UndertaleObject
     {
-        public UndertaleObservableList<int> InstanceIDs { get; set; } = new();
+        public UndertaleObservableList<uint> InstanceIDs { get; set; } = new(8);
 
         public void Serialize(UndertaleWriter writer)
         {
@@ -2792,7 +2839,7 @@ public class UndertaleRoom : UndertaleNamedResource, INotifyPropertyChanged, IDi
             InstanceIDs.SetCapacity(count);
             for (int i = 0; i < count; i++)
             {
-                InstanceIDs.InternalAdd(reader.ReadInt32());
+                InstanceIDs.InternalAdd(reader.ReadUInt32());
             }
         }
     }
