@@ -6,6 +6,8 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using UndertaleModLib.Util;
+using UndertaleModLib.Project;
+using UndertaleModLib.Project.SerializableAssets;
 
 namespace UndertaleModLib.Models;
 
@@ -18,7 +20,7 @@ public enum AnimSpeedType : uint
 /// <summary>
 /// Sprite entry in the data file.
 /// </summary>
-public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyPropertyChanged, IDisposable
+public class UndertaleSprite : UndertaleNamedResource, IProjectAsset, PrePaddedObject, INotifyPropertyChanged, IDisposable
 {
     /// <summary>
     /// The name of the sprite.
@@ -204,6 +206,49 @@ public class UndertaleSprite : UndertaleNamedResource, PrePaddedObject, INotifyP
         VectorShapes = null;
         VectorCollisionMaskRLEData = null;
         VectorFrameToShapeMap = null;
+    }
+
+    /// <inheritdoc/>
+    ISerializableProjectAsset IProjectAsset.GenerateSerializableProjectAsset(ProjectContext projectContext)
+    {
+        SerializableSprite serializable = new();
+        serializable.PopulateFromData(projectContext, this);
+        return serializable;
+    }
+
+    /// <inheritdoc/>
+    public string ProjectName => Name?.Content ?? "<unknown name>";
+
+    /// <inheritdoc/>
+    public SerializableAssetType ProjectAssetType => SerializableAssetType.Sprite;
+
+    /// <inheritdoc/>
+    public bool ProjectExportable
+    {
+        get
+        {
+            if (Name?.Content is null)
+            {
+                return false;
+            }
+            if (IsSpecialType && SSpriteType is not SpriteType.Normal) // TODO: support more types
+            {
+                return false;
+            }
+            if ((Textures?.Count ?? 0) == 0)
+            {
+                return false;
+            }
+            if (Textures.Any(t => t?.Texture is null))
+            {
+                return false;
+            }
+            if (CollisionMasks.Any(c => c?.Data is null))
+            {
+                return false;
+            }
+            return true;
+        }
     }
 
     /// <summary>
