@@ -215,6 +215,17 @@ namespace UndertaleModTool.Windows
             isSearchInProgress = false;
         }
 
+        string GetCodeString(UndertaleCode code)
+        {
+            // First, try to retrieve source from project (if available)
+            if (mainWindow.Project is null || !mainWindow.Project.TryGetCodeSource(code, out string decompiled))
+            {
+                // Source isn't available - perform decompile
+                decompiled = new Underanalyzer.Decompiler.DecompileContext(decompileContext, code, mainWindow.Data.ToolInfo.DecompilerSettings).DecompileToString();
+            }
+            return decompiled;
+        }
+
         private void SearchInUndertaleCode(UndertaleCode code)
         {
             try
@@ -223,8 +234,7 @@ namespace UndertaleModTool.Windows
                 {
                     var codeText = isInAssembly
                         ? code.Disassemble(mainWindow.Data.Variables, mainWindow.Data.CodeLocals?.For(code), mainWindow.Data.CodeLocals is null)
-                        : TryGetProfileModeGML(code.Name.Content)
-                            ?? new Underanalyzer.Decompiler.DecompileContext(decompileContext, code, mainWindow.Data.ToolInfo.DecompilerSettings).DecompileToString();
+                        : GetCodeString(code);
                     SearchInCodeText(code.Name.Content, codeText);
                 }
                 
@@ -237,17 +247,6 @@ namespace UndertaleModTool.Windows
 
             Interlocked.Increment(ref progressCount);
             Dispatcher.Invoke(() => loaderDialog.ReportProgress(progressCount));
-        }
-
-        static string TryGetProfileModeGML(string codeName)
-        {
-            if (SettingsWindow.ProfileModeEnabled && mainWindow.ProfileHash is not null)
-            {
-                string path = Path.Join(Settings.ProfilesFolder, mainWindow.ProfileHash, "Temp", codeName + ".gml");
-                if (File.Exists(path))
-                    return File.ReadAllText(path).Replace("\r\n", "\n");
-            }
-            return null;
         }
 
         private void SearchInCodeText(string codeName, string codeText)
